fix: multiple exports from files
