+++ src/app/(mobile-ui)/points/invites/page.tsx
@@ -1,7 +1,8 @@
 'use client'
 
 import PageContainer from '@/components/0_Bruddle/PageContainer'
-import Card, { getCardPosition } from '@/components/Global/Card'
+import Card from '@/components/Global/Card'
+import { getCardPosition } from '@/components/Global/Card/card.utils'
 import NavHeader from '@/components/Global/NavHeader'
 import PeanutLoading from '@/components/Global/PeanutLoading'
 import TransactionAvatarBadge from '@/components/TransactionDetails/TransactionAvatarBadge'
