+++ src/components/Badges/index.tsx
@@ -5,7 +5,7 @@ import type { StaticImageData } from 'next/image'
 import NavHeader from '../Global/NavHeader'
 import { useRouter } from 'next/navigation'
 import { getBadgeIcon } from './badge.utils'
-import { getCardPosition } from '../Global/Card'
+import { getCardPosition } from '../Global/Card/card.utils'
 import EmptyState from '../Global/EmptyStates/EmptyState'
 import { Icon } from '../Global/Icons/Icon'
 import ActionModal from '../Global/ActionModal'
