+++ src/components/Common/CountryList.tsx
@@ -9,7 +9,7 @@ import EmptyState from '@/components/Global/EmptyStates/EmptyState'
 import { SearchInput } from '@/components/SearchInput'
 import Image from 'next/image'
 import { useMemo, useState, useDeferredValue, type ReactNode } from 'react'
-import { getCardPosition } from '../Global/Card'
+import { getCardPosition } from '../Global/Card/card.utils'
 import { useGeoLocation } from '@/hooks/useGeoLocation'
 import { CountryListSkeleton } from './CountryListSkeleton'
 import AvatarWithBadge from '../Profile/AvatarWithBadge'
