+++ src/components/Common/CountryListSkeleton.tsx
@@ -1,6 +1,6 @@
 'use client'
 import { ActionListCard } from '../ActionListCard'
-import { getCardPosition } from '../Global/Card'
+import { getCardPosition } from '../Global/Card/card.utils'
 
 /**
  * Displays a country list skeleton!
