+++ src/components/Kyc/KYCStatusDrawerItem.tsx
@@ -1,0 +1,15 @@
+import Card from '@/components/Global/Card'
+import StatusBadge, { type StatusType } from '../Global/Badges/StatusBadge'
+import { KYCStatusIcon } from './KYCStatusIcon'
+
+export const KYCStatusDrawerItem = ({ status }: { status: StatusType }) => {
+    return (
+        <Card position="single" className="flex items-center gap-4">
+            <KYCStatusIcon />
+            <div className="flex flex-col gap-2">
+                <h3 className="text-lg font-extrabold">Identity verification</h3>
+                <StatusBadge status={status} className="w-fit" size="small" />
+            </div>
+        </Card>
+    )
+}
