+++ src/components/Kyc/KYCStatusIcon.tsx
@@ -1,0 +1,5 @@
+import AvatarWithBadge from '../Profile/AvatarWithBadge'
+
+export const KYCStatusIcon = () => {
+    return <AvatarWithBadge icon="user-id" className="bg-yellow-1" size="extra-small" />
+}
