+++ src/components/Kyc/states/KycProcessing.tsx
@@ -1,5 +1,5 @@
 import { PaymentInfoRow } from '@/components/Payment/PaymentInfoRow'
-import { KYCStatusDrawerItem } from '../KycStatusItem'
+import { KYCStatusDrawerItem } from '../KYCStatusDrawerItem'
 import Card from '@/components/Global/Card'
 import { useMemo } from 'react'
 import { formatDate } from '@/utils/general.utils'
