+++ src/components/Profile/views/RegionsVerification.view.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { ActionListCard } from '@/components/ActionListCard'
-import { getCardPosition } from '@/components/Global/Card'
+import { getCardPosition } from '@/components/Global/Card/card.utils'
 import EmptyState from '@/components/Global/EmptyStates/EmptyState'
 import { Icon } from '@/components/Global/Icons/Icon'
 import NavHeader from '@/components/Global/NavHeader'
