+++ src/hooks/useGraphPreferences.ts
@@ -5,7 +5,7 @@ import type {
     ForceConfig,
     VisibilityConfig,
     ExternalNodesConfig,
-} from '@/components/Global/InvitesGraph'
+} from '@/components/Global/InvitesGraph/types'
 
 const GRAPH_PREFS_KEY = 'invite-graph-preferences'
